feat:modify diamond.js, equilateral.js,validation.js,permission.js and add index.js files
