+++ extra/utils/index.js
@@ -1,0 +1,2 @@
+export { default as validateUser } from "./validation";
+export { default as hasPermission } from "./permissions";
