+++ extra/utils/permissions.js
@@ -1,19 +1,11 @@
-let permissions = {
-  getUsers: {
-    all: ["head-trainer"],
-    read: ["trainee", "trainer"],
-    write: ["trainer"],
-    delete: []
-  },
-  writeUser: {
-    all: ["trainer"],
-    read: ["trainee", "trainer"],
-    delete: []
-  }
-};
-function hasPermission(moduleName, role, permissionType) {
+import { permissions } from "../constants";
+
+export default function hasPermission(moduleName, role, permissionType) {
   if (permissions[moduleName]) {
-    if (permissions[moduleName][permissionType].includes(role) || permissions[moduleName]["all"].includes(role)) {
+    if (
+      permissions[moduleName][permissionType].includes(role) ||
+      permissions[moduleName]["all"].includes(role)
+    ) {
       console.log("true");
     } else {
       console.log("false");
@@ -22,8 +14,3 @@ function hasPermission(moduleName, role, permissionType) {
     console.log("false");
   }
 }
-hasPermission("getUsers", "trainee", "read");
-hasPermission("get", "trainee", "read");
-hasPermission("getUsers", "trainee", "all");
-hasPermission("getUsers", "head-trainer", "read");
-hasPermission("writeUser", "trainee", "read");
