+++ extra/utils/validation.js
@@ -1,22 +1,6 @@
-const users = [
-  {
-    traineeEmail: "trainee1@successive.tech",
-    reviewerEmail: "reviewer1@successive.tech"
-  },
-  {
-    traineeEmail: "trainee1@gmail.com",
-    reviewerEmail: "reviewer1@gmail.com"
-  },
-  {
-    traineeEmail: "trainee2@xyz.tech",
-    reviewerEmail: "reviewer2@xyz.tech"
-  }
-];
-function validateEmail(email) {
-  let regex = /^([A-Za-z0-9_\-\.])+\@(successive.tech)$/;
-  return regex.test(email);
-}
-function validateUsers(users) {
+import { validateEmail } from "./helpers";
+
+export default function validateUsers(users) {
   const validUser = [];
   const invalidUser = [];
   let validCount = 0,
@@ -35,4 +19,3 @@ function validateUsers(users) {
   console.log(`${validCount} valid users are ${validUser}`);
   console.log(`${invalidCount} invalid users are ${invalidUser}`);
 }
-validateUsers(users);
